scrapers/requirements.txt
# Snitched.ai Scraper Dependencies
# Install: pip install -r requirements.txt

# Social Media Scrapers
facebook-scraper>=0.2.59
selenium>=4.15.0
webdriver-manager>=4.0.1

# FEC API
requests>=2.31.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9
supabase>=2.3.4

# Data Processing
pandas>=2.1.0
python-dateutil>=2.8.2

# Sentiment Analysis
textblob>=0.17.1

# OpenPlanter (OSINT framework)
# Note: Install separately via pip install -e ./OpenPlanter

# Utilities
tqdm>=4.66.1
pytz>=2023.3
